Move UploadObserver out, simplify ID handling
