+++ android/src/main/java/com/bluechilli/flutteruploader/MethodCallHandlerImpl.java
@@ -1,30 +1,20 @@
 package com.bluechilli.flutteruploader;
 
 import android.content.Context;
-import android.text.TextUtils;
-import android.util.Log;
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.lifecycle.Observer;
 import androidx.work.BackoffPolicy;
 import androidx.work.Constraints;
 import androidx.work.Data;
 import androidx.work.NetworkType;
 import androidx.work.OneTimeWorkRequest;
-import androidx.work.WorkInfo;
 import androidx.work.WorkManager;
 import androidx.work.WorkRequest;
 import com.bluechilli.flutteruploader.plugin.StatusListener;
 import com.google.gson.Gson;
-import com.google.gson.reflect.TypeToken;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
 import io.flutter.plugin.common.MethodChannel.Result;
-import java.io.BufferedReader;
-import java.io.FileReader;
-import java.lang.ref.WeakReference;
-import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -35,16 +25,14 @@ import java.util.concurrent.TimeUnit;
 
 public class MethodCallHandlerImpl implements MethodCallHandler {
 
-  /** Plugin registration. */
-  private static final String TAG = "flutter_upload_task";
+  /** The generic {@link WorkManager} tag which matches any upload. */
+  public static final String FLUTTER_UPLOAD_WORK_TAG = "flutter_upload_task";
 
   private final Context context;
+
   private int connectionTimeout;
 
   @NonNull private final StatusListener statusListener;
-
-  private Gson gson = new Gson();
-  private int taskIdKey = 0;
 
   private static final List<String> VALID_HTTP_METHODS = Arrays.asList("POST", "PUT", "PATCH");
 
@@ -53,99 +41,6 @@ public class MethodCallHandlerImpl implements MethodCallHandler {
     this.connectionTimeout = timeout;
     this.statusListener = listener;
   }
-
-  static class UploadObserver implements Observer<List<WorkInfo>> {
-    private final WeakReference<MethodCallHandlerImpl> plugin;
-
-    UploadObserver(MethodCallHandlerImpl plugin) {
-      this.plugin = new WeakReference<>(plugin);
-    }
-
-    @Override
-    public void onChanged(List<WorkInfo> workInfoList) {
-      MethodCallHandlerImpl plugin = this.plugin.get();
-
-      if (plugin == null) {
-        return;
-      }
-
-      for (WorkInfo info : workInfoList) {
-        String id = info.getId().toString();
-
-        switch (info.getState()) {
-          case RUNNING:
-            {
-              Data progress = info.getProgress();
-              Log.d(TAG, "progress update: " + progress.toString());
-              Log.d(TAG, "progress update tag: " + info.getTags());
-              plugin.statusListener.onUpdateProgress(
-                  info.getId().toString(),
-                  progress.getInt("status", -1),
-                  progress.getInt("progress", -1));
-            }
-            break;
-          case FAILED:
-            {
-              final Data outputData = info.getOutputData();
-              int failedStatus = outputData.getInt(UploadWorker.EXTRA_STATUS, UploadStatus.FAILED);
-              int statusCode = outputData.getInt(UploadWorker.EXTRA_STATUS_CODE, 500);
-              String code = outputData.getString(UploadWorker.EXTRA_ERROR_CODE);
-              String errorMessage = outputData.getString(UploadWorker.EXTRA_ERROR_MESSAGE);
-              String[] details = outputData.getStringArray(UploadWorker.EXTRA_ERROR_DETAILS);
-
-              plugin.statusListener.onFailed(
-                  id, failedStatus, statusCode, code, errorMessage, details);
-            }
-            break;
-          case CANCELLED:
-            plugin.statusListener.onFailed(
-                id, UploadStatus.CANCELED, 500, "flutter_upload_cancelled", null, null);
-            break;
-          case SUCCEEDED:
-            {
-              final Data outputData = info.getOutputData();
-              int status = outputData.getInt(UploadWorker.EXTRA_STATUS, UploadStatus.COMPLETE);
-              int statusCode = outputData.getInt(UploadWorker.EXTRA_STATUS_CODE, 500);
-              Map<String, String> headers = null;
-              Type type = new TypeToken<Map<String, String>>() {}.getType();
-              String headerJson = outputData.getString(UploadWorker.EXTRA_HEADERS);
-              if (headerJson != null) {
-                headers = plugin.gson.fromJson(headerJson, type);
-              }
-
-              String response = extractResponse(outputData);
-              plugin.statusListener.onCompleted(id, status, statusCode, response, headers);
-            }
-            break;
-        }
-      }
-    }
-
-    String extractResponse(Data outputData) {
-      String response = outputData.getString(UploadWorker.EXTRA_RESPONSE);
-      if (TextUtils.isEmpty(response)) {
-        String responseFile = outputData.getString(UploadWorker.EXTRA_RESPONSE_FILE);
-        if (!TextUtils.isEmpty(responseFile)) {
-          StringBuilder buffer = new StringBuilder();
-
-          try (BufferedReader br = new BufferedReader(new FileReader(responseFile))) {
-            String st;
-            while ((st = br.readLine()) != null) {
-              buffer.append(st);
-            }
-            response = buffer.toString();
-
-          } catch (Throwable ignored) {
-            response = "";
-          }
-        }
-      }
-
-      return response;
-    }
-  }
-
-  @Nullable private UploadObserver uploadObserver;
 
   @Override
   public void onMethodCall(MethodCall call, @NonNull Result result) {
@@ -183,20 +78,6 @@ public class MethodCallHandlerImpl implements MethodCallHandler {
     result.success(null);
   }
 
-  void startObservers() {
-    uploadObserver = new UploadObserver(this);
-    WorkManager.getInstance(context).getWorkInfosByTagLiveData(TAG).observeForever(uploadObserver);
-  }
-
-  void stopObservers() {
-    if (uploadObserver != null) {
-      WorkManager.getInstance(context)
-          .getWorkInfosByTagLiveData(TAG)
-          .removeObserver(uploadObserver);
-      uploadObserver = null;
-    }
-  }
-
   private void enqueue(MethodCall call, MethodChannel.Result result) {
     String url = call.argument("url");
     String method = call.argument("method");
@@ -219,8 +100,6 @@ public class MethodCallHandlerImpl implements MethodCallHandler {
       return;
     }
 
-    taskIdKey++;
-
     List<FileItem> items = new ArrayList<>();
 
     for (Map<String, String> file : files) {
@@ -229,8 +108,7 @@ public class MethodCallHandlerImpl implements MethodCallHandler {
 
     WorkRequest request =
         buildRequest(
-            new UploadTask(
-                taskIdKey, url, method, items, headers, parameters, connectionTimeout, false, tag));
+            new UploadTask(url, method, items, headers, parameters, connectionTimeout, false, tag));
     WorkManager.getInstance(context).enqueue(request);
     String taskId = request.getId().toString();
     result.success(taskId);
@@ -258,12 +136,9 @@ public class MethodCallHandlerImpl implements MethodCallHandler {
       return;
     }
 
-    taskIdKey++;
-
     WorkRequest request =
         buildRequest(
             new UploadTask(
-                taskIdKey,
                 url,
                 method,
                 Collections.singletonList(FileItem.fromJson(files)),
@@ -286,7 +161,7 @@ public class MethodCallHandlerImpl implements MethodCallHandler {
   }
 
   private void cancelAll(MethodCall call, MethodChannel.Result result) {
-    WorkManager.getInstance(context).cancelAllWorkByTag(TAG);
+    WorkManager.getInstance(context).cancelAllWorkByTag(FLUTTER_UPLOAD_WORK_TAG);
     result.success(null);
   }
 
@@ -304,8 +179,7 @@ public class MethodCallHandlerImpl implements MethodCallHandler {
             .putString(UploadWorker.ARG_METHOD, task.getMethod())
             .putInt(UploadWorker.ARG_REQUEST_TIMEOUT, task.getTimeout())
             .putBoolean(UploadWorker.ARG_BINARY_UPLOAD, task.isBinaryUpload())
-            .putString(UploadWorker.ARG_UPLOAD_REQUEST_TAG, task.getTag())
-            .putInt(UploadWorker.ARG_ID, task.getId());
+            .putString(UploadWorker.ARG_UPLOAD_REQUEST_TAG, task.getTag());
 
     List<FileItem> files = task.getFiles();
 
@@ -325,7 +199,7 @@ public class MethodCallHandlerImpl implements MethodCallHandler {
     return new OneTimeWorkRequest.Builder(UploadWorker.class)
         .setConstraints(
             new Constraints.Builder().setRequiredNetworkType(NetworkType.CONNECTED).build())
-        .addTag(TAG)
+        .addTag(FLUTTER_UPLOAD_WORK_TAG)
         .setBackoffCriteria(BackoffPolicy.EXPONENTIAL, 5, TimeUnit.SECONDS)
         .setInputData(dataBuilder.build())
         .build();
