+++ android/src/main/java/com/bluechilli/flutteruploader/UploadTask.java
@@ -14,19 +14,16 @@ public class UploadTask {
   private int _requestTimeoutInSeconds;
   private boolean _binaryUpload;
   private String _tag;
-  private int _id;
 
   public UploadTask(
-      int id,
-      String url,
-      String method,
-      List<FileItem> files,
-      Map<String, String> headers,
-      Map<String, String> data,
-      int requestTimeoutInSeconds,
-      boolean binaryUpload,
-      String tag) {
-    _id = id;
+			String url,
+			String method,
+			List<FileItem> files,
+			Map<String, String> headers,
+			Map<String, String> data,
+			int requestTimeoutInSeconds,
+			boolean binaryUpload,
+			String tag) {
     _url = url;
     _method = method;
     _files = files;
@@ -71,9 +68,5 @@ public class UploadTask {
 
   public String getTag() {
     return _tag;
-  }
-
-  public int getId() {
-    return _id;
   }
 }
